Kill window after a temporary split
This is mainly encountered when using embark to show options in a new
split window. This now works well in that case, and removes the window.
I had cases in mind before where I might want that window to stick around,
but I don't remember what they were, so we'll see if those come up at
some point now that this delete is here.